+++ core/FrontController.php
@@ -64,6 +64,9 @@ class FrontController extends Singleton
     const DEFAULT_LOGIN = 'anonymous';
     const DEFAULT_TOKEN_AUTH = 'anonymous';
 
+    // public for tests
+    public static $requestId = null;
+
     /**
      * Set to false and the Front Controller will not dispatch the request
      *
@@ -109,6 +112,10 @@ class FrontController extends Singleton
             'file' => $e->getFile(),
             'line' => $e->getLine(),
         );
+
+        if (isset(self::$requestId)) {
+            $error['request_id'] = self::$requestId;
+        }
 
         $error['backtrace'] = ' on ' . $error['file'] . '(' . $error['line'] . ")\n";
         $error['backtrace'] .= $e->getTraceAsString();
@@ -240,6 +247,11 @@ class FrontController extends Singleton
     public static function triggerSafeModeWhenError()
     {
         $lastError = error_get_last();
+
+        if (!empty($lastError) && isset(self::$requestId)) {
+            $lastError['request_id'] = self::$requestId;
+        }
+
         if (!empty($lastError) && $lastError['type'] == E_ERROR) {
             $lastError['backtrace'] = ' on ' . $lastError['file'] . '(' . $lastError['line'] . ")\n"
                 . ErrorHandler::getFatalErrorPartialBacktrace();
@@ -265,6 +277,8 @@ class FrontController extends Singleton
         if ($this->initialized) {
             return;
         }
+
+        self::setRequestIdHeader();
 
         $this->initialized = true;
 
@@ -677,5 +691,19 @@ class FrontController extends Singleton
         }
 
         return $authAdapter;
+    }
+
+    public static function getUniqueRequestId()
+    {
+        if (self::$requestId === null) {
+            self::$requestId = substr(Common::generateUniqId(), 0, 5);
+        }
+        return self::$requestId;
+    }
+
+    private static function setRequestIdHeader()
+    {
+        $requestId = self::getUniqueRequestId();
+        Common::sendHeader("X-Matomo-Request-Id: $requestId");
     }
 }